+++ db/migrate/20240306112000_add_confirm_date_academy_transferred_attributes_to_transfer.rb
@@ -1,0 +1,5 @@
+class AddConfirmDateAcademyTransferredAttributesToTransfer < ActiveRecord::Migration[7.0]
+  def change
+    add_column :transfer_tasks_data, :confirm_date_academy_transferred_date_transferred, :date
+  end
+end